requirements.txt
# ── Core ──────────────────────────────────────────────────
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3
langgraph==0.2.28

# ── Vector Store & Retrieval ──────────────────────────────
chromadb==0.5.5
rank-bm25==0.2.2
sentence-transformers==3.1.1   # cross-encoder reranker (ms-marco-MiniLM)

# ── Document Processing ───────────────────────────────────
pymupdf==1.24.10               # PDF parsing (fitz)
Pillow==10.4.0                 # image pre-processing

# ── Web & API ─────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.9        # file uploads

# ── Search Fallback ───────────────────────────────────────
duckduckgo-search==6.2.13
tavily-python==0.5.0           # optional; needs TAVILY_API_KEY

# ── Sandbox ───────────────────────────────────────────────
RestrictedPython==7.4

# ── UI ────────────────────────────────────────────────────
streamlit==1.38.0

# ── Observability ─────────────────────────────────────────
langsmith==0.1.129             # optional; needs LANGSMITH_API_KEY
python-dotenv==1.0.1

# ── Dev & Test ────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
